# api/openapi.yaml
openapi: 3.1.0
info:
  title: Jasper API
  version: 0.2.0
  description: |
    Jasper HTTP API: notes CRUD keyed by ULID, the file-tree projection,
    admin endpoints (migration status, full re-index) backing DATA-10, and
    tags, backlinks and wiki-link search-titles
    (TAGS-03, TAGS-04, TAGS-06, TAGS-07, LINKS-06, LINKS-08).

    Source of truth: every route Jasper exposes lives in this file (API-01).
    Adding a route here without a Go handler causes a compile error
    under oapi-codegen StrictServerInterface (API-02).
servers:
  - url: /api/v1
    description: Same-origin (Go binary or Vite dev proxy)
...
paths:
...
  /files:
    post:
      tags: [notes]
      operationId: createFile
      summary: Upload an OS file into the notes/ tree at the specified directory.
      description: |
        Multipart upload landing under <dataDir>/notes/<path>. Used by the
        sidebar tree's OS-file drop target — drop on a folder row → file lands
        inside that folder; drop on a note/file → file lands as sibling; drop
        on empty area → file lands at vault root.

        path is passed as a query parameter (mirroring DELETE /folders and the
        sibling GET /files endpoint) because OpenAPI 3.1 has no
        native multi-segment path-wildcard syntax and oapi-codegen does not
        emit chi `*` catch-all routes. Co-locating GET + POST at /files keeps
        the surface coherent: same wire shape for both verbs.

        Refuses .md uploads (markdown bodies must be created via POST /notes).
        Reuses generateUniqueFilename for collision-safe naming.
        Enforces the 100 MB upload cap. Same 5-rule path-traversal
        pipeline as GET /files plus a Lstat that ensures the
        target is a directory (NO auto-mkdir — the user creates folders via
        the existing tree UI before dropping).
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: |
            Target DIRECTORY (relative to notes/). "" = vault root.
            "gallery" = notes/gallery/. "gallery/sub" = notes/gallery/sub/.
            Must not contain `..` segments or absolute path prefixes;
            must point to an existing directory (returns 400 otherwise).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded; returns relative path + final filename.
          content:
            application/json:
              schema:
                type: object
                required: [path, name, size_bytes]
                properties:
                  path:
                    type: string
                    description: Relative path under notes/ (e.g. "gallery/photo-1.png").
                  name:
                    type: string
                    description: Final filename after collision-safe rename.
                  size_bytes:
                    type: integer
                    format: int64
                  content_type:
                    type: string
                    description: MIME sniff of the uploaded bytes.
        "400":
          description: Invalid path / filename / target dir does not exist / .md upload refused.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Symlink rejected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: File too large (>100 MB).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [notes]
      operationId: getFile
      summary: Stream any file under the notes vault by relative path.
      description: |
        Generic file streamer. Returns ANY file located under <dataDir>/notes/<path>
        without requiring a note context. Used by the sidebar tree's file-click
        handler when the user selects a non-markdown file (image attachments,
        PDFs, etc.). Returns 404 for .md files (use /notes/{id} instead).

        path is passed as a query parameter (mirroring DELETE /folders) because
        OpenAPI 3.1 has no native multi-segment path-wildcard syntax and
        oapi-codegen does not emit chi `*` catch-all routes. The query-param
        approach handles slash-containing relative paths cleanly while keeping
        the route inside the generated StrictServerInterface.

        **The generated handler for this route is a stub and is unreachable.**
        app.go and lifecycle.go mount `api.Server.ServeFile` AFTER
        `api.HandlerFromMux`, so it wins under chi's last-registration-wins —
        the same arrangement `/ws` uses. The path stays in this spec so every
        HTTP route the server answers is described here.

        The hand-mounted handler exists because the generated response type can
        set only Content-Type and Content-Length, and hard-codes the former to
        `application/octet-stream`. The live route needs neither value:

        - **Content-Type is sniffed** via http.DetectContentType, with an
          explicit `image/svg+xml` override for `.svg` — DetectContentType
          returns `text/xml` for SVG, which browsers refuse to render in `<img>`.
        - **Content-Disposition** is `inline` for an allowlist of media types
          (images including SVG, audio, video, PDF) and `attachment` for
          everything else, so a planted `.html` or `.xhtml` cannot be navigated
          to as a live document.
        - **Content-Security-Policy** is `default-src 'none'; sandbox` on every
          response, alongside `X-Content-Type-Options: nosniff`. Files are
          served from the app's own origin, so without this an SVG containing
          `<script>`, opened as a top-level document, would execute with full
          same-origin API access to every note. `sandbox` loads it into an
          opaque origin with scripting disabled; it does not affect subresource
          loads, so images embedded in notes still render.

        Containment: the path is resolved through `fsstore.ResolveContained`,
        which decides containment before stat'ing the leaf, so a path escaping
        the vault via a symlinked ancestor is rejected identically whether or
        not the target exists.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: |
            Canonical relative path under notes/ (e.g. "attachments/photo.png",
            "gallery/attachments/note.pdf"). Must not contain `..` segments or
            absolute path prefixes (validated server-side).
      responses:
        "200":
          description: |
            Binary stream. The declared `application/octet-stream` reflects the
            generated stub, NOT the live handler, which sends a sniffed media
            type plus Content-Disposition and the raw-file security headers
            described above. oapi-codegen has no way to express "whatever the
            handler decides"; treat the description as authoritative here.
          content:
            application/octet-stream: {}
        "400":
          description: Invalid path (path traversal attempt or absolute path).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Symlink rejected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: File not found OR path is a .md file (use /notes/{id}).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [notes]
      operationId: deleteFile
      summary: Delete a non-markdown file under notes/.
      description: |
        Mirrors GetFile's path-traversal pipeline. Refuses .md (those are notes —
        use DELETE /notes/{id}). Refuses directories (use DELETE /folders). Single
        os.Remove call once the path resolves under notes/.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Canonical relative path under notes/ (e.g. "attachments/photo.png").
      responses:
        "204":
          description: File deleted.
        "400":
          description: Invalid path / .md path refused / target is a directory.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Symlink rejected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: File not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /files/move:
    post:
      tags: [notes]
      operationId: postFileMove
      summary: Rename / move a non-markdown file under notes/.
      description: |
        Mirrors POST /folders/move for non-markdown files. Both src_path and
        dst_path are relative under notes/; both run through the 5-rule
        path-traversal pipeline. Refuses .md (those go through POST /notes/{id}/move).
        Refuses overwrite (409 if dst already exists). Atomic os.Rename on POSIX
        same-filesystem.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [src_path, dst_path]
              properties:
                src_path:
                  type: string
                  description: Existing relative path under notes/.
                dst_path:
                  type: string
                  description: Target relative path under notes/.
      responses:
        "200":
          description: Move committed; returns the new path.
          content:
            application/json:
              schema:
                type: object
                required: [path, name]
                properties:
                  path:
                    type: string
                  name:
                    type: string
        "400":
          description: Invalid src/dst path / .md path refused.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Symlink rejected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Source file not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Destination already exists (refuses overwrite).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "not_found"
        message:
          type: string
          example: "no note with id 00000000000000000000000001"
...